NoteApp/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
git-filter-repo==2.47.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2